Update lti-tool-learndash.php
When plugin is active LearnDash stops recording data to the learndash_user_activity table. Found issue was in the lti_tool_learndash_update_activity function fixed by adding return with args to function. This allows LearnDash activity to be saved both to the local site and be pushed to LTi with no issues.